done
Сделал верстку для мобилки. Добавил тени на тогл радио-кнопку, убрал в календаре возможность выбирать дату меньше текущего дня.
