+++ src/components/HomeTab/HomeTabLarge.jsx
@@ -1,64 +1,78 @@
+import { useSelector } from 'react-redux';
+import transactionSelectors from '../../redux/transactions/transactions-selectors';
+import ModalAddTransaction from '../ModalAddTransaction';
+import ButtonAddTransaction from '../ButtonAddTransaction';
 import './HomeTabLarge.scss';
 
 function HomeTabLarge({ transactions }) {
+  const isModalAddTransactionOpen = useSelector(
+    transactionSelectors.getIsModalAddTransactionOpen,
+  );
+
   const ItemSpendClass = 'home-table__cell home-table__cell_spend';
   const ItemIncomeClass = 'home-table__cell home-table__cell_income';
 
   return (
-    <table className="home-table">
-      <thead className="home-table__header">
-        <tr className="home-table__row">
-          <th className="home-table__cell">Дата</th>
-          <th className="home-table__cell">Тип</th>
-          <th className="home-table__cell">Категория</th>
-          <th className="home-table__cell">Комментарий</th>
-          <th className="home-table__cell">Сумма</th>
-          <th className="home-table__cell">Баланс</th>
-        </tr>
-      </thead>
-      <tbody className="home-table__body">
-        {transactions.map(({ _id, type, date, money, category, comment }) => (
-          <tr key={_id} className="home-table__row">
-            <td className="home-table__cell">{date}</td>
-            <td className="home-table__cell">{type}</td>
-            <td className="home-table__cell">{category.name}</td>
-            <td className="home-table__cell">{comment}</td>
-            <td className={type === '-' ? ItemSpendClass : ItemIncomeClass}>
-              {money}
-            </td>
-            <td className="home-table__cell">6 9000.00</td>
+    <>
+      {isModalAddTransactionOpen && <ModalAddTransaction />}
+      <table className="home-table">
+        <thead className="home-table__header">
+          <tr className="home-table__row">
+            <th className="home-table__cell">Дата</th>
+            <th className="home-table__cell">Тип</th>
+            <th className="home-table__cell">Категория</th>
+            <th className="home-table__cell">Комментарий</th>
+            <th className="home-table__cell">Сумма</th>
+            <th className="home-table__cell">Баланс</th>
           </tr>
-        ))}
+        </thead>
+        <tbody className="home-table__body">
+          {transactions.map(({ _id, type, date, money, category, comment }) => (
+            <tr key={_id} className="home-table__row">
+              <td className="home-table__cell">{date}</td>
+              <td className="home-table__cell">{type}</td>
+              <td className="home-table__cell">{category.name}</td>
+              <td className="home-table__cell">{comment}</td>
+              <td className={type === '-' ? ItemSpendClass : ItemIncomeClass}>
+                {money}
+              </td>
+              <td className="home-table__cell">6 9000.00</td>
+            </tr>
+          ))}
 
-        {/* !!!!!!!!!!!!!STATIC DATA!!!!!!!!!!!! */}
-        <tr className="home-table__row">
-          <td className="home-table__cell">05.01.19</td>
-          <td className="home-table__cell">+</td>
-          <td className="home-table__cell">Регулярный доход</td>
-          <td className="home-table__cell">
-            Бонус за январь TEST LAAAAARGE MESSSAGGGEEEE
-          </td>
-          <td className="home-table__cell home-table__cell_income">8 000.00</td>
-          <td className="home-table__cell">14 900.00</td>
-        </tr>
-        <tr className="home-table__row">
-          <td className="home-table__cell">07.01.19</td>
-          <td className="home-table__cell">-</td>
-          <td className="home-table__cell">Машина</td>
-          <td className="home-table__cell">Масло</td>
-          <td className="home-table__cell home-table__cell_spend">1000.00</td>
-          <td className="home-table__cell">13 900.00</td>
-        </tr>
-        <tr className="home-table__row">
-          <td className="home-table__cell">07.01.19</td>
-          <td className="home-table__cell">-</td>
-          <td className="home-table__cell">Продукты</td>
-          <td className="home-table__cell">Овощи на неделю</td>
-          <td className="home-table__cell home-table__cell_spend">280.00</td>
-          <td className="home-table__cell">13 870.00</td>
-        </tr>
-      </tbody>
-    </table>
+          {/* !!!!!!!!!!!!!STATIC DATA!!!!!!!!!!!! */}
+          <tr className="home-table__row">
+            <td className="home-table__cell">05.01.19</td>
+            <td className="home-table__cell">+</td>
+            <td className="home-table__cell">Регулярный доход</td>
+            <td className="home-table__cell">
+              Бонус за январь TEST LAAAAARGE MESSSAGGGEEEE
+            </td>
+            <td className="home-table__cell home-table__cell_income">
+              8 000.00
+            </td>
+            <td className="home-table__cell">14 900.00</td>
+          </tr>
+          <tr className="home-table__row">
+            <td className="home-table__cell">07.01.19</td>
+            <td className="home-table__cell">-</td>
+            <td className="home-table__cell">Машина</td>
+            <td className="home-table__cell">Масло</td>
+            <td className="home-table__cell home-table__cell_spend">1000.00</td>
+            <td className="home-table__cell">13 900.00</td>
+          </tr>
+          <tr className="home-table__row">
+            <td className="home-table__cell">07.01.19</td>
+            <td className="home-table__cell">-</td>
+            <td className="home-table__cell">Продукты</td>
+            <td className="home-table__cell">Овощи на неделю</td>
+            <td className="home-table__cell home-table__cell_spend">280.00</td>
+            <td className="home-table__cell">13 870.00</td>
+          </tr>
+        </tbody>
+      </table>
+      <ButtonAddTransaction />
+    </>
   );
 }
