+++ src/components/Modal/Modal.scss
@@ -15,7 +15,7 @@
   left: 50%;
   transform: translate(-50%, -50%);
   width: 100%;
-  padding: 40px 60px 65px 65px;
+  padding: 40px 65px 60px 65px;
   background-color: $modal-color;
   border-radius: 20px;
 }
