+++ src/components/ModalAddTransaction/AddTransaction/AddTransaction.scss
@@ -1,0 +1,213 @@
+@import 'styles/shared';
+
+@media screen and (min-width: 320px) {
+  .transaction {
+    &__title {
+      font-family: $secondary-font;
+      @include font(24px, 36px);
+      font-weight: 400;
+      text-align: center;
+      margin-bottom: 40px;
+    }
+
+    &__wrapper,
+    &__form {
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+    }
+
+    &__button {
+      display: inline-block;
+      width: 300px;
+      height: 50px;
+      border-radius: 20px;
+      cursor: pointer;
+      text-transform: uppercase;
+      @include font(18px, 26px);
+      font-weight: 400;
+      border: none;
+      background-color: $button-color;
+      color: $secondary-text-color;
+      margin-bottom: 20px;
+    }
+
+    &__button--cancel {
+      background-color: inherit;
+      border: 1px solid $accent-color;
+      color: $accent-color;
+      margin-bottom: 0;
+    }
+  }
+
+  .input {
+    outline: none;
+    border: none;
+    border-bottom: 1px solid $placeholder-color;
+    cursor: pointer;
+
+    &:hover {
+      border-color: $accent-color;
+    }
+
+    &:focus {
+      border-color: $accent-color;
+    }
+  }
+
+  .form {
+    &__checkbox {
+      margin-bottom: 40px;
+    }
+
+    &__wrapper {
+      display: flex;
+      flex-direction: column;
+      margin-bottom: 40px;
+      position: relative;
+    }
+
+    &__sum {
+      margin-bottom: 40px;
+    }
+
+    &__text {
+      margin-bottom: 50px;
+    }
+  }
+
+  .text__input {
+    width: 280px;
+    height: 74px;
+    @include font(18px, 26px);
+    font-weight: 400;
+    padding-left: 20px;
+    padding-bottom: 50px;
+
+    &::placeholder {
+      color: $placeholder-color;
+    }
+  }
+
+  .sum__input {
+    width: 280px;
+    height: 34px;
+    @include font(18px, 27px);
+    font-weight: 400;
+    padding-left: 20px;
+
+    &::placeholder {
+      color: $placeholder-color;
+      padding-bottom: 10px;
+    }
+  }
+
+  .date {
+    &__input {
+      width: 280px;
+      height: 34px;
+      padding-left: 20px;
+      font-weight: 400;
+      @include font(18px, 27px);
+      color: $text-color;
+    }
+
+    &__icon {
+      width: 24px;
+      height: 24px;
+      fill: $accent-color;
+      position: absolute;
+      right: 20px;
+      top: 4px;
+      pointer-events: none;
+    }
+  }
+
+  .select {
+    &__wrapper {
+      position: relative;
+    }
+
+    &__icon {
+      position: absolute;
+      right: 5px;
+      top: 15px;
+      z-index: 9;
+      width: 18px;
+      height: 9px;
+    }
+  }
+
+  .checkbox {
+    &__wrapper {
+      margin-bottom: 40px;
+      display: flex;
+      align-items: center;
+    }
+
+    &__span {
+      font-weight: 700;
+      @include font(16px, 24px);
+      color: $placeholder-border-color;
+    }
+
+    &__button {
+      margin: 0 20px;
+    }
+  }
+
+  .active-income {
+    color: $income-color;
+  }
+
+  .active-spend {
+    color: $spend-color;
+  }
+
+  .switch-icon__spend {
+    width: 20px;
+    position: absolute;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+  }
+
+  .react-switch-bg {
+    border: 1px solid $placeholder-border-color;
+  }
+}
+
+@media screen and (min-width: 768px) {
+  .transaction {
+    &__title {
+      @include font(30px, 45px);
+    }
+  }
+
+  .form {
+    &__wrapper {
+      flex-direction: row;
+    }
+
+    &__sum {
+      margin-right: 30px;
+    }
+  }
+
+  .text__input {
+    width: 408px;
+    height: 34px;
+    padding-bottom: 10px;
+  }
+
+  .sum__input {
+    width: 190px;
+    padding-left: 75px;
+  }
+
+  .date {
+    &__input {
+      width: 190px;
+    }
+  }
+}
