+++ src/components/ModalAddTransaction/ModalAddTransaction.scss
@@ -1,55 +1,15 @@
 @import 'styles/shared';
 
 .modal {
-  position: relative;
   max-height: 617px;
-  width: 540px;
+  max-width: 540px;
+  padding: 40px 65px 60px 65px;
 }
 
 .transaction {
   &__title {
-    font-family: $secondary-font;
     @include font(30px, 45px);
-    font-weight: 400;
-    text-align: center;
-    margin-bottom: 40px;
   }
-
-  &__wrapper {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-
-  &__form {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-  }
-
-  &-button {
-    display: inline-block;
-    width: 300px;
-    height: 50px;
-    border-radius: 20px;
-    cursor: pointer;
-    text-transform: uppercase;
-    @include font(18px, 26px);
-    font-weight: 400;
-  }
-}
-
-.transaction-button__add {
-  border: none;
-  background-color: $button-color;
-  color: $secondary-text-color;
-  margin-bottom: 20px;
-}
-
-.transaction-button__cancel {
-  background-color: inherit;
-  border: 1px solid $accent-color;
-  color: $accent-color;
 }
 
 .close-modal {
@@ -74,145 +34,4 @@
     left: 50%;
     transform: translate(-50%, -50%);
   }
-}
-
-.form {
-  &__checkbox {
-    margin-bottom: 40px;
-  }
-
-  &__wrapper {
-    display: flex;
-    margin-bottom: 40px;
-    position: relative;
-  }
-
-  &__sum {
-    margin-right: 30px;
-  }
-
-  &__text {
-    margin-bottom: 50px;
-  }
-}
-
-.input {
-  outline: none;
-  border: none;
-  border-bottom: 1px solid $placeholder-color;
-  cursor: pointer;
-
-  &:hover {
-    border-color: $accent-color;
-  }
-
-  &:focus {
-    border-color: $accent-color;
-  }
-}
-
-.text__input {
-  width: 408px;
-  height: 34px;
-  @include font(18px, 26px);
-  font-weight: 400;
-  padding-left: 20px;
-
-  &::placeholder {
-    color: $placeholder-color;
-    padding-bottom: 10px;
-  }
-}
-
-.sum__input {
-  width: 190px;
-  height: 34px;
-  @include font(18px, 26px);
-  font-weight: 400;
-  padding-left: 75px;
-
-  &::placeholder {
-    color: $placeholder-color;
-    padding-bottom: 10px;
-  }
-}
-
-.date {
-  &__input {
-    width: 190px;
-    height: 34px;
-    padding-left: 20px;
-    font-weight: 400;
-    @include font(18px, 27px);
-    color: $text-color;
-  }
-
-  &__icon {
-    width: 24px;
-    height: 24px;
-    fill: $accent-color;
-    position: absolute;
-    right: 20px;
-    top: 4px;
-    pointer-events: none;
-  }
-}
-
-.select {
-  &__wrapper {
-    position: relative;
-  }
-
-  &__icon {
-    position: absolute;
-    right: 5px;
-    top: 15px;
-    z-index: 9;
-    width: 18px;
-    height: 9px;
-  }
-}
-
-.checkbox {
-  &__wrapper {
-    margin-bottom: 40px;
-    display: flex;
-    align-items: center;
-  }
-
-  &__span {
-    font-weight: 700;
-    @include font(16px, 24px);
-    color: $placeholder-border-color;
-  }
-
-  &__button {
-    margin-right: 20px;
-  }
-}
-
-.span {
-  &-income {
-    margin-right: 20px;
-  }
-}
-
-.active-income {
-  color: $income-color;
-}
-
-.active-spend {
-  color: $spend-color;
-}
-
-.switch-icon__spend {
-  width: 20px;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
-
-.react-switch-bg {
-  border: 1px solid $placeholder-border-color;
 }
